% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KCIT.R
\name{KCIT}
\alias{KCIT}
\title{RCIT and RCoT - tests whether x and y are conditionally independent given z. Calls U_KCI if z is empty.}
\usage{
KCIT(x, y, z, Bootstrap = TRUE)
}
\arguments{
\item{x}{Random variable x.}

\item{y}{Random variable y.}

\item{z}{Random variable z.}

\item{Bootstrap}{Estimate the null by bootstrap. Default is TRUE. Runs slightly faster when set to FALSE but with less accurate p-values.}
}
\value{
The p-value \code{p}
}
\description{
RCIT and RCoT - tests whether x and y are conditionally independent given z. Calls U_KCI if z is empty.
}

