% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix2.R
\name{matrix2}
\alias{matrix2}
\title{Converts a vector to a matrix, or does nothing if the input is a matrix.}
\usage{
matrix2(mat)
}
\arguments{
\item{mat}{Vector or matrix.}
}
\value{
Matrix.
}
\description{
Converts a vector to a matrix, or does nothing if the input is a matrix.
}
\examples{
matrix2(rnorm(10));
matrix2(matrix(rnorm(10,2),ncol=2));
}

