% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalizes each column of a matrix to mean zero unit variance.}
\usage{
normalize(mat)
}
\arguments{
\item{mat}{Matrix}
}
\value{
A matrix where each column has mean zero and unit variance.
}
\description{
Normalizes each column of a matrix to mean zero unit variance.
}
\examples{
normalize(matrix(rnorm(10,2),ncol=2))
}

