% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcit-package.R
\docType{package}
\name{rcit-package}
\alias{rcit}
\alias{rcit-package}
\title{rcit: Causal Inference Tools for R}
\description{
CIT defines functions useful for causal inference.
}
\author{
\strong{Maintainer}: Guilherme Salome \email{guilhermesalome@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
