% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_binary_treatment.R
\name{t_test_binary_treatment}
\alias{t_test_binary_treatment}
\title{Difference in means test}
\usage{
t_test_binary_treatment(data, treatment, ignore_cols)
}
\arguments{
\item{data}{Tibble}

\item{treatment}{Column with binary treatment indicator}

\item{ignore_cols}{String vector with columns to be ignored in the test}
}
\description{
Test if averages from treatment and control groups are statistically equal.
}
\examples{
t_test_binary_treatment(data, "treatment", "output")
}
\keyword{difference-in-means,}
\keyword{t-test}
