% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U_KCI.R
\name{U_KCI}
\alias{U_KCI}
\title{RCIT and RCoT - tests whether x and y are unconditionally independent..}
\usage{
U_KCI(x, y)
}
\arguments{
\item{x}{Random variable x.}

\item{y}{Random variable y.}
}
\value{
The p-value \code{p}
}
\description{
RCIT and RCoT - tests whether x and y are unconditionally independent..
}

